requirements.txt
dune-client
python-dotenv
pandas
requests
matplotlib
python-telegram-bot==20.7
pytest
flake8
